Attempt at refactoring the save functionality into app. Issue with continuous rerenders creating noticable slowdown
