+++ packages/editor/src/components/SavedBrowser.tsx
@@ -1,9 +1,6 @@
-import { debounce } from "lodash";
-import { useCallback, useEffect, useRef } from "react";
-import { useRecoilCallback, useRecoilState, useRecoilValue } from "recoil";
+import { useRecoilValue } from "recoil";
 import { v4 as uuid } from "uuid";
 import {
-  autosaveTimerState,
   currentAppUser,
   currentWorkspaceState,
   savedFilesState,
@@ -12,106 +9,105 @@ import {
   useDeleteWorkspace,
   useLoadLocalWorkspace,
   useSaveNewWorkspace,
-  useSaveWorkspace,
 } from "../state/callbacks.js";
 import { logInWrapper } from "../utils/firebaseUtils.js";
 import BlueButton from "./BlueButton.js";
 import FileButton from "./FileButton.js";
 
-/**
- * Allows user to save workspace with ctrl+s or cmd+s
- * Due to current layout scheme, this runs no matter what side tab the user
- * is on. If layout scheme ever changes, this may need to be moved to App
- */
-const saveShortcutHook = () => {
-  const saveWorkspace = useSaveWorkspace();
+// /**
+//  * Allows user to save workspace with ctrl+s or cmd+s
+//  * Due to current layout scheme, this runs no matter what side tab the user
+//  * is on. If layout scheme ever changes, this may need to be moved to App
+//  */
+// const saveShortcutHook = () => {
+//   const saveWorkspace = useSaveWorkspace();
 
-  // Need useRecoilCallback for snapshot, otherwise currentWorkspace outdated
-  const handleShortcut = useRecoilCallback(
-    ({ snapshot }) =>
-      async (event: KeyboardEvent) => {
-        const currentWorkspace = snapshot.getLoadable(
-          currentWorkspaceState,
-        ).contents;
-        if (event.repeat) return;
-        // Cmd+s or Cntrl+s
-        if (
-          (event.metaKey || event.ctrlKey) &&
-          event.key === "s" &&
-          currentWorkspace.metadata.location.kind == "stored" &&
-          !currentWorkspace.metadata.location.saved
-        ) {
-          event.preventDefault();
+//   // Need useRecoilCallback for snapshot, otherwise currentWorkspace outdated
+//   const handleShortcut = useRecoilCallback(
+//     ({ snapshot }) =>
+//       async (event: KeyboardEvent) => {
+//         const currentWorkspace = snapshot.getLoadable(
+//           currentWorkspaceState,
+//         ).contents;
+//         if (event.repeat) return;
+//         // Cmd+s or Ctrl+s
+//         if (
+//           (event.metaKey || event.ctrlKey) &&
+//           event.key === "s" &&
+//           currentWorkspace.metadata.location.kind == "stored" &&
+//           !currentWorkspace.metadata.location.saved
+//         ) {
+//           event.preventDefault();
 
-          saveWorkspace();
-        }
-      },
-  );
+//           saveWorkspace();
+//         }
+//       },
+//   );
 
-  useEffect(() => {
-    document.addEventListener("keydown", handleShortcut, {
-      passive: false,
-    });
+//   useEffect(() => {
+//     document.addEventListener("keydown", handleShortcut, {
+//       passive: false,
+//     });
 
-    // Cleanup
-    return () => document.removeEventListener("keydown", handleShortcut);
-  }, []);
-};
+//     // Cleanup
+//     return () => document.removeEventListener("keydown", handleShortcut);
+//   }, []);
+// };
 
-/**
- * Autosaves every 5 seconds after a user has finished editing
- * Known bug: The first edit to a diagram after load will not trigger this.
- * This is due to nuances with currentWorkspace state not updating in time.
- * Due to current layout scheme, this runs no matter what side tab the user
- * is on. If layout scheme ever changes, this may need to be moved to App
- */
-const autosaveHook = () => {
-  const currentWorkspace = useRecoilValue(currentWorkspaceState);
-  const isInitialRender = useRef(true);
-  const saveWorkspace = useSaveWorkspace();
-  const [autosaveTimerValue, autosaveTimerSetter] =
-    useRecoilState(autosaveTimerState);
+// /**
+//  * Autosaves every 5 seconds after a user has finished editing
+//  * Known bug: The first edit to a diagram after load will not trigger this.
+//  * This is due to nuances with currentWorkspace state not updating in time.
+//  * Due to current layout scheme, this runs no matter what side tab the user
+//  * is on. If layout scheme ever changes, this may need to be moved to App
+//  */
+// const autosaveHook = () => {
+//   const currentWorkspace = useRecoilValue(currentWorkspaceState);
+//   const isInitialRender = useRef(true);
+//   const saveWorkspace = useSaveWorkspace();
+//   const [autosaveTimerValue, autosaveTimerSetter] =
+//     useRecoilState(autosaveTimerState);
 
-  /**
-   * useCallback necessary for debounce to work. Without debounce, every
-   * character entered will trigger the function.
-   */
-  const autosaveLogic = useCallback(
-    debounce(async () => {
-      console.log("hit outer");
-      // Reset autosave timer
-      if (autosaveTimerValue != null) {
-        clearTimeout(autosaveTimerValue);
-      }
-      // Set new timer, after 5 seconds have elapsed without edit
-      const newTimeoutId = setTimeout(() => {
-        console.log("hit inner");
-        console.log(currentWorkspace.metadata.location);
-        if (
-          currentWorkspace.metadata.location.kind == "stored"
-          // !currentWorkspace.metadata.location.saved
-        ) {
-          saveWorkspace();
-        }
-      }, 3000);
-      autosaveTimerSetter(newTimeoutId);
-    }, 500),
-    // So that updates to these values won't be reflected in execution
-    [autosaveTimerValue, currentWorkspace.metadata],
-  );
+//   /**
+//    * useCallback necessary for debounce to work. Without debounce, every
+//    * character entered will trigger the function.
+//    */
+//   const autosaveLogic = useCallback(
+//     debounce(async () => {
+//       console.log("hit outer");
+//       // Reset autosave timer
+//       if (autosaveTimerValue != null) {
+//         clearTimeout(autosaveTimerValue);
+//       }
+//       // Set new timer, after 5 seconds have elapsed without edit
+//       const newTimeoutId = setTimeout(() => {
+//         console.log("hit inner");
+//         console.log(currentWorkspace.metadata.location);
+//         if (
+//           currentWorkspace.metadata.location.kind == "stored"
+//           // !currentWorkspace.metadata.location.saved
+//         ) {
+//           saveWorkspace();
+//         }
+//       }, 3000);
+//       autosaveTimerSetter(newTimeoutId);
+//     }, 500),
+//     // So that updates to these values won't be reflected in execution
+//     [autosaveTimerValue, currentWorkspace.metadata],
+//   );
 
-  useEffect(() => {
-    if (isInitialRender.current) {
-      isInitialRender.current = false;
-      return;
-    }
-    autosaveLogic();
-  }, [
-    currentWorkspace.files.substance.contents,
-    currentWorkspace.files.style.contents,
-    currentWorkspace.files.domain.contents,
-  ]);
-};
+//   useEffect(() => {
+//     if (isInitialRender.current) {
+//       isInitialRender.current = false;
+//       return;
+//     }
+//     autosaveLogic();
+//   }, [
+//     currentWorkspace.files.substance.contents,
+//     currentWorkspace.files.style.contents,
+//     currentWorkspace.files.domain.contents,
+//   ]);
+// };
 
 export default function SavedFilesBrowser() {
   const savedFiles = useRecoilValue(savedFilesState);
@@ -122,8 +118,8 @@ export default function SavedFilesBrowser() {
   const useLogin = logInWrapper();
   const currentUser = useRecoilValue(currentAppUser);
 
-  saveShortcutHook();
-  autosaveHook();
+  // saveShortcutHook();
+  // autosaveHook();
 
   return (
     <>
